adding syr2k tests, swapping syr2k and syrkx gtest files, small test fixes.
